requirements.txt
aiohttp==3.5.4

PyYAML~=5.3.1
colorlog~=4.6.2
SQLAlchemy~=1.3.22